+++ .clangd
@@ -1,0 +1,5 @@
+# https://clangd.llvm.org/config
+CompileFlags:
+  Add: [
+    "-Wno-c++11-compat-deprecated-writable-strings"
+  ]
