+++ .vscode/settings.json
@@ -15,7 +15,7 @@
     "files.insertFinalNewline": true,
     "files.trimFinalNewlines": true,
     "files.associations": {
-        "*.inc": "c",
+        "*.inc": "cpp",
         ".clangd": "yaml"
     },
     // Disable C/C++ IntelliSense, use clangd instead
