+++ include/d/flag/tboxflag_manager.h
@@ -1,0 +1,33 @@
+#ifndef D_FLAG_TBOXFLAG_MANAGER_H
+#define D_FLAG_TBOXFLAG_MANAGER_H
+
+#include "d/flag/bitwise_flag_helper.h"
+#include "d/flag/committable_flag_manager.h"
+#include "d/flag/flag_space.h"
+
+class TBoxFlagManager : public CommittableFlagManager {
+public:
+    FlagSpace mFlagSpace;
+    u16 mSceneIndex;
+    BitwiseFlagHelper mFlagHelper;
+
+    static u16 sTBoxFlags[2];
+
+    static TBoxFlagManager *sInstance;
+
+    virtual void doCommit() override;
+    bool checkUncommittedFlag(u16 flag);
+    TBoxFlagManager();
+    virtual ~TBoxFlagManager() {}
+    void init();
+    void copyFromSave(u32 sceneIndex);
+    bool checkFlag(u16 sceneIndex, u16 flag);
+    virtual u16 getFlagCount() const;
+    void setFlag(u16 flag);
+    bool checkUncommittedFlag(u16 sceneIndex, u16 flag);
+    u16 checkUncommittedFlag2(u16 flag) {
+        return checkUncommittedFlag(flag);
+    }
+};
+
+#endif
