+++ src/REL/d/t/d_t_touch.cpp
@@ -2,7 +2,7 @@
 
 #include "c/c_math.h"
 #include "toBeSorted/area_math.h"
-#include "toBeSorted/sceneflag_manager.h"
+#include "d/flag/sceneflag_manager.h"
 
 
 SPECIAL_ACTOR_PROFILE(TOUCH_TAG, dTgTouchTag, fProfile::TOUCH_TAG, 0x028E, 0, 0);
