+++ src/d/a/d_a_base.cpp
@@ -8,7 +8,7 @@
 #include "toBeSorted/event.h"
 #include "toBeSorted/event_manager.h"
 #include "toBeSorted/file_manager.h"
-#include "toBeSorted/misc_flag_managers.h"
+#include "d/flag/enemyflag_manager.h"
 #include "toBeSorted/room_manager.h"
 #include "toBeSorted/scgame.h"
 #include "toBeSorted/special_item_drop_mgr.h"
