+++ src/d/flag/committable_flag_manager.inc
@@ -1,0 +1,12 @@
+#include "d/flag/committable_flag_manager.h"
+
+/* 0x800BE7B0 */
+bool CommittableFlagManager::commitIfNecessary() {
+    if (mNeedsCommit) {
+        doCommit();
+        mNeedsCommit = false;
+        return true;
+    } else {
+        return false;
+    }
+}
