+++ src/d/flag/unk_flag_stuff.inc
@@ -1,4 +1,4 @@
-#include "toBeSorted/unk_flag_stuff.h"
+#include "d/flag/unk_flag_stuff.h"
 
 u16 UnkFlagStuff::calculateMask(s32 shift) {
     return (2 << (shift & 0xF)) - 1;
